+++ src/lib.rs
@@ -21,9 +21,11 @@ mod low_level;
 #[cfg(test)]
 mod tests;
 
-pub use containers::{ArrayLike, BoxedArrayLike, SliceLike};
+pub use containers::{ArrayLike, BorrowedSliceLike, BoxedArrayLike, SliceLike};
 pub use encoding::Encoding;
-pub use low_level::{TryFromArray, TryFromSliceRef};
+
+#[cfg(feature = "generic-array-014")]
+pub use containers::GenericArray014;
 
 // Specifically enable `Hex` for tests, since we need some encoding to be specified.
 // Should be removed when https://github.com/rust-lang/cargo/issues/2911 is fixed.